Prevent looping steps from invalidating themselves
When a fork on a step forms a loop - that is it returns to previously visited step to enter repeated values - then don't invalidate that section of the journey when it is not being followed.

When a loop is being followed then the steps which form the loop should be invalidated so that once the loop has been commenced then it must be completed.