Fix tests broken in previous PR
One of the functional tests had been broken by a previous PR, that had accidentally included an `it.only`.

Remove the `only` and fix the broke test.

Fix is that a loop should only be considered to be "started" when a step has actually been submitted. Going to a page an dnot submitting it should allow return. Add an extra step into the loop and submit the first step of the loop to test this case.